% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowsum2.R
\name{rowsum2}
\alias{rowsum2}
\title{Flexible summation of matrix/data.frame rows}
\usage{
rowsum2(data, x = c("na", "zero", "sum"), op, value, pattern, ...)
}
\arguments{
\item{data}{\code{data.frame} or \code{matrix}.}

\item{x}{\code{string} method, one of either: \code{"na"}, \code{"zero"}, or \code{"sum"}.}

\item{op}{Optional \code{string} describing operator to use for comparing row sums
to a desired value. One of either: \code{"=="}, \code{"!="}, \code{"<="}, \code{">="}, \code{"<"}, \code{">"}.}

\item{value}{Optional \code{numeric} value to compare the row sums against.}

\item{pattern}{Optional \code{string} containing a regular expression to match
column names via \code{\link[base]{grep}}.}

\item{...}{Other arguments to be passed into \code{\link[base]{grep}}.}
}
\value{
Returns a \code{logical} vector if 'op' and 'value' have been specified.
Otherwise returns a \code{numeric} vector.
}
\description{
Count the sum of values, sum of zeros, or sum of \code{NA}s in the
each row of a data.frame or matrix. Optionally count only certain columns
and/or compare the row sums to a desired value.
}
\examples{
mat <- matrix(c(NA, 1:6, 0), nrow = 4, ncol = 2,
              dimnames = list(NULL, c("sample1", "sample2")))
df <- data.frame(mat)

# works with data.frame or matrix
rowsum2(mat, "na")
rowsum2(df, "na")

# get just the sums (should use rowSums directly in this particular case),
# or sum of na/zero for each row
rowsum2(mat, "sum")
rowsum2(mat, "na")
rowsum2(mat, "zero")

# compare the row sums against a value of interest to
# see which rows 'pass'
rowsum2(mat, "sum", ">", 5)

# get row sums for only certain columns based on
# matching columns with grep
rowsum2(mat, "na", pattern = "sample1")

# combine row sum comparison and column matching
rowsum2(df, "na", "==", 0, "sample1")

}
