% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_x.R
\name{filter_x}
\alias{filter_x}
\title{Flexible filtering of matrix/data.frames}
\usage{
filter_x(
  data,
  x = c("na", "zero", "sum"),
  op = c("==", "!=", "<=", ">=", "<", ">"),
  value,
  pattern,
  ...,
  setop
)
}
\arguments{
\item{data}{\code{data.frame} or \code{matrix}.}

\item{x}{\code{string} summation method, one of either: \code{"na"}, \code{"zero"}, or \code{"sum"}.}

\item{op}{\code{string} or \code{character} vector describing operator to use for
comparing row sums to a desired value.
Possible values: \code{"=="}, \code{"!="}, \code{"<="}, \code{">="}, \code{"<"}, \code{">"}.}

\item{value}{\code{numeric} value or vector to compare the row sums against.}

\item{pattern}{Optional \code{string} or \code{character} vector containing regular
expression(s) to match column names via \code{\link[base]{grep}}.}

\item{...}{Other arguments to be passed into \code{\link[base]{grep}}.}

\item{setop}{Optional \code{string} to indicate the method to combine the rows
returned by multiple filtering steps. This must be specified if using a
vector for \code{op}, \code{value}, or \code{pattern}.}
}
\value{
Returns \code{data} with the rows not matching the specified condition(s)
removed.
}
\description{
Filter a matrix or data.frame based on the sum of values, zero,
or \code{NA} in each row. Optionally filter based on only certain columns.
Also optionally perform multiple filtering steps on different sets of columns,
wherein the function is vectorised over \code{op}, and \code{value}, and \code{pattern}
(see examples).
}
\examples{
# need to add some examples
}
